Revised `graphviz` dataflow-rendering support in `borrowck::mir`.
A future commit will relink this to the parent module code; (as of
this commit it is dead code, apart from the `mod graphviz` declaration
to ensure it builds).

The main difference is that this code is now parameterized over the
`BitDenotation`, so that it can be used to render dataflow results
independent of how the dataflow bitvectors are interpreted; see where
reference to `MoveOut` is replaced by the type parameter `D`.